requirements.txt
asgiref==3.2.10
astroid==2.4.2
auditlog3==1.0.1
autopep8==1.5.4
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.1
django-admin-interface==0.14.0
django-allauth==0.43.0
django-colorfield==0.3.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-polymorphic==3.0.0
django-property-filter==0.8.0
django-rest-auth==0.9.5
django-rest-polymorphic==0.1.9
django-tempus-dominus==5.1.2.13
django-vueformgenerator==0.2.3
djangorestframework==3.11.1
drf-generators==0.5.0
idna==2.10
isort==5.5.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.1
oauthlib==3.1.0
Pillow==7.2.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
wrapt==1.12.1
